requirements.txt
openai>=1.12.0  # For API access and embeddings
faiss-cpu>=1.7.4  # For similarity search (use faiss-gpu if GPU is available)
numpy>=1.24.0    # Required by faiss
typing>=3.7.4.3  # For type hints
python-dotenv>=1.0.0  # For managing environment variables (optional but recommended)
uuid>=1.30       # For generating unique IDs
logging>=0.5.1.2 # For logging functionality
datetime>=5.4    # For handling timestamps
dotenv>=0.19.1   # For managing environment variables